refactor: remove dead code and simplify serialization
- Remove unused VOICE_CONNECTION_TIMEOUT_MS constant and its re-export
- Remove unnecessary export from SerializedSong (internal type)
- Inline buildQueuedSong into its two call sites in player.ts
- Move formatSong from socket.ts to new serialization.ts module
