+++ packages/api/src/routes/songs.ts
@@ -5,7 +5,8 @@ import { db } from '../lib/db';
 import { getUserDisplayName } from '../lib/displayName';
 import { json } from '../lib/json';
 
-import { emitSongAdded, emitSongDeleted, emitSongUpdated, formatSong } from '../lib/socket';
+import { emitSongAdded, emitSongDeleted, emitSongUpdated } from '../lib/socket';
+import { formatSong } from '../lib/serialization';
 import {
   clampMaxVideos,
   fetchPlaylistMetadata,
