--- packages/bot/src/lib/constants.ts
@@ -1,2 +1,0 @@
-/** Timeout (ms) when waiting for voice connection to be ready */
-export const VOICE_CONNECTION_TIMEOUT_MS = 5_000;
